+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - Sources: 'Sources\Select playlist(s)' submenu being disabled when it shouldn't.
 - Shortcuts: fixed bug on original script related to usage of 'Ctrl + A' (select All) shortcut crashing the panel while using Playlist sources if 'Behaviour\Mode\Keystroke' was set to 'Send to playlist'. When set to 'Select' it simply didn't work. Now the shortcut works as intended.
 - Drag n' drop: workaround for CUI/SMP drag n' drop bug while dragging tracks directly from album list panel, without library view playlist usage. See [here](https://hydrogenaudio.org/index.php/topic,116669.msg1077838.html#msg1077838) and [CUI Issue 459](https://github.com/reupen/album_list_panel/issues/459).
+- Statistics: fixed bitrate not working.
 
 ## [1.0.0] - 2026-01-30
 ### Added